1.0 plottwatt API Complete
This is a major update. The functions are all working as far as I know.
The example sketches can be used to list the meters, add meters and
delete meters. Once you have the meters set up correctly, you can update
the information in basicPlottWatt.ino to make it work for your account.