Definitions in INO
Up to this point the main definitions have been in the INO file and
passed to the other functions. The duemilinove spelling? can only do 42
samples before it runs out of memory. I'm going to change the
definitions to a header file and not pass them to test efficiency
improvements.